+++ lib/src/eval/compiler/expression/expression.dart
@@ -20,6 +20,7 @@ import 'package:dart_eval/src/eval/compiler/expression/parenthesized.dart';
 import 'package:dart_eval/src/eval/compiler/expression/postfix.dart';
 import 'package:dart_eval/src/eval/compiler/expression/prefix.dart';
 import 'package:dart_eval/src/eval/compiler/expression/property_access.dart';
+import 'package:dart_eval/src/eval/compiler/expression/rethrow.dart';
 import 'package:dart_eval/src/eval/compiler/expression/throw.dart';
 import 'package:dart_eval/src/eval/compiler/reference.dart';
 import 'package:dart_eval/src/eval/compiler/type.dart';
@@ -69,6 +70,8 @@ Variable compileExpression(Expression e, CompilerContext ctx,
     return compileCascadeExpression(e, ctx);
   } else if (e is AsExpression) {
     return compileAsExpression(e, ctx);
+  } else if (e is RethrowExpression) {
+    return compileRethrowExpression(ctx, e);
   }
 
   throw CompileError('Unknown expression type ${e.runtimeType}');