+++ CHANGELOG.md
@@ -20,7 +20,8 @@
 - Support for null-shorted method calls and property access
 - Fix type resolution with self-referential generic types such as
   `class T implements Comparable<T>`
-- Add binding for `Exception`
+- Support for `rethrow`
+- Add binding for `Exception` and `RangeError`
 - Runtime errors now print out the scope stack for easier debugging
 - Reformat line length to 80 for standardization with Dart ecosystem
 - Documentation improvements